requirements.txt
# ── LangChain Core Stack ───────────────────────────────────────────────────────
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0

# ── Vector Store ───────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Local Embeddings (no external API) ────────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.2.0
transformers>=4.40.0

# ── LLM Backend ───────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Document Processing ────────────────────────────────────────────────────────
pypdf>=4.2.0

# ── UI ─────────────────────────────────────────────────────────────────────────
gradio>=4.44.0,<5.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
requests>=2.32.0
numpy>=1.26.0
python-dotenv>=1.0.1
